+++ pandapower/control/controller/trafo/ContinuousTapControl.py
@@ -4,8 +4,8 @@
 # and Energy System Technology (IEE), Kassel. All rights reserved.
 
 import numpy as np
+from pandapower.auxiliary import read_from_net, write_to_net
 from pandapower.control.controller.trafo_control import TrafoController
-from pandapower.toolbox_elm_selection import read_from_net, write_to_net
 
 
 class ContinuousTapControl(TrafoController):
