+++ pandapower/control/controller/trafo/DiscreteTapControl.py
@@ -3,8 +3,8 @@
 # Copyright (c) 2016-2023 by University of Kassel and Fraunhofer Institute for Energy Economics
 # and Energy System Technology (IEE), Kassel. All rights reserved.
 import numpy as np
+from pandapower.auxiliary import read_from_net, write_to_net
 from pandapower.control.controller.trafo_control import TrafoController
-from pandapower.toolbox_elm_selection import read_from_net, write_to_net
 
 class DiscreteTapControl(TrafoController):
     """
